[ci,nightly] Disable use of cached test results in nightly runs
If FPGA test results can be cached through a remote cache, then we want
to ensure that all nightly test runs run without this caching (for the
test runs specifically, not the rest of the build), to allow us to be
able to spot flaky tests when they occur.

This commit achieves this by adding the `--cache_test_results=no`
argument to all invocations of `./bazelisk.sh test` in the nightly
Github Actions workflow.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>