[bazel] Hotswap deprecated `sync` command in bazelisk.sh
This command has been deprecated when using Bzlmod. It only works with
`WORKSPACE` and hybrid mode. For some reason Bazel completely crashes
with a JVM exception when you try to use it rather than printing a nice
warning.

This commit prints a warning that `sync` is deprecated and swaps the
command for `fetch` which has almost identical arguments.

This is a temporary measure to prevent breaking contributors who have
git post-checkout hooks installed which run `bazel sync --configure`.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>