[rom_ctrl,dv] Tidy up fatal alert tracking in kmac error check vseq
This sequence sends an error message from KMAC and checks that it does
indeed get reflected as an alert from rom_ctrl. This works fine, but
the sequence followed up with a dut_init call, which is a bit strange!

I suspect this was to avoid getting to cip_base_vseq::post_start with
an alert in progress. This was a problem because
rom_ctrl_kmac_err_chk_vseq didn't set the expect_fatal_alerts flag.

Do so, and drop the extra reset.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>