[orchestrator] fix bug in computation of Device ID
This fixes a bug in the computation and construction of the Device ID
when parts are provisioned from a TestLocked* stage, i.e., if part has
already previously run through CP and then runs through FT at a later
stage. In this scenario, the there is no CP device ID returned to the
Host (since CP stage is not run), so FT must trust what already exists
in flash, and the final device ID is read out by the host over JTAG at
the end of provisioning.

Signed-off-by: Tim Trippel <ttrippel@google.com>