[autogen_testutils] Change how IPs with DIFs are detected
In the next commit, we will remove all the autogenerated files in
sw/device/lib/dif/autogen. This will break how autogen_testutils
detects IPs with diff. The logic is change to the following: go
through the list of IPs for the top of interest, check in
sw/device/lib/dif if there is a file called dif_<ip>.h. This produces
exactly the same result.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>