[bazel] update rules_rust to 0.56
This version splits bindgen to a separate extension, so rename
accordingly.

Updating rules_rust_bindgen pulls in a new version of bindgen which
requires a higher Rust version then what we currently use. So update
Rust to 1.83, which is the default version used in rules_rust 0.56.
Similarly, the nightly version is updated to 2024-11-28 which is the
default nightly for rules_rust 0.56.

The new version of clippy starts to warn about needless lifetimes in
cases where it's necessary or idiomatic to keep (e.g. trait
implementations), so disable it for now.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>