[hw,tlul_adapter_reg,rtl] Do not gate a_ready with a_valid
This creates a combinational path between the input and output TLUL
port. This can be pretty long and impacts the synthesis in high-frequency
designs.

Originally, the busy_i signal was gated with a_valid. But I think that
does not make sense. The busy signal is an external input and is in no
relation to TLUL. In fact, the only usage of that signal is for shadowed
resets, to keep a_ready low until all resets are released.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>