[hw,tlul_adapter_reg,rtl] Do not gate a_ready with a_valid
This creates a combinational path between the input and output TLUL
port. This can be pretty long and impacts the synthesis in high-frequency
designs.

For some cases where CDC is involved, busy depends on a_valid. For these
cases, factoring in a_valid is done in the reg_top, explicitly for when
needed. That cuts the path for cases where only the fast main clock is
involved. For CDC cases, there is still this combinational path. But
because that is typically using the way slower AON domain, this is not
that critical there.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>