[kmac, sival] Keymgr-state-agnostic `kmac_sideload_functest`
This commit fixes the two `kmac_sideload_functest` tests that were
failing due to a misconfigured key version parameter. Some ROMs set
the max key version in the keymgr to 0, which causes sideloading to
abort if any other version value is being programmed. Note that the
proper functioning of valid/invalid keymgr version numbers is part of
the `chip_sw_keymgr_derive_{sealing, attestation}` tests.

This bug has been mentioned as part of #21706.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>