[rom_ctrl, dv] Assertion coverage for rom_ctrl
There was some reported assertion coverage hole for rom_ctrl.
All the assertions were basically related and they haven't
seen the execution of them on a particular event. They should
happen when we run a rom_ctrl_sec_cm test and inject errors
in rom_ctrl's adapter fifos. Based on that an alert should
fire, which was true. However, when we inject anything bad
to the fifos, it fails some assertions in rom_ctrl and
rom_tlul_assert_device. So, instead of turning all the
assertions off for rom_ctrl (in rom_ctrl_commom_vseq), we
just turn the affected assertions off. This resolves the
assertion hole for rom_ctrl

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>