[manuf] enable alerts during individualization
This updates the individualization firmware to enable alerts and capture
crashdump information if requested by the host. This may be used to
debug alert escalations that may occur during various manufacturing
operations such as OTP programming. This feature is piped through the
orchestrator script to enable turning it on or off as needed during
bringup / debugging efforts.

Signed-off-by: Tim Trippel <ttrippel@google.com>