[sca,aes] Get rid of aes_sca_error_t
When porting the AES SCA tests from the old simpleserial to the new
uJSON framework, we also copied the aes_sca_error_t structure. This
structure was used to indicate whether a function inside the aes_sca
program returned with an error.

Instead, we simply can use the default status_t with the TRY()
construct. This (a) simplifies the code and (b) follows the
coding style of the OpenTitan code base.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 1af8f1c)