[pentest] Load fixed seed after each dif_aes_start
With the recent changes in the PRNG, it is necessary to reseed the
internal AES state after the AUX register has been set by using the
dif_aes_start() function. We need to do this when we want to switch
off masking for AES SCA tests.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 45d184c)