[hw,otp_ctrl,rtl] Parametrize the DFT config and response port
This parametrization allows the usage of prim_otp with different
DFT config and response ports. One usage is for example when using
macro tiling within the prim instance and have different instances
with different configuration in the design. By passing the port
type as a parameter we don't need to pull in all the tiling related
parameters into the design.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>