[rom] Make the DEV key valid in TEST and RMA
Make the DEV key valid in TEST and RMA. Because of the limited number of
key slots in OTP, we allow the DEV key to be valid in DEV, TEST and RMA.
This will allow a single set of keys (e.g. 3xPROD + 1xDEV) to be viable
for devices provisioned to PROD and DEV mission mode states while still
allowing the use of the DEV key for TEST and RMA states.

Signed-off-by: Chris Frantz <cfrantz@google.com>
(cherry picked from commit 35ec68c1b83c9c2cf93c13072cac28d647fa6098)