[rom] fix bug in ROM that computes OTP measurements
This fixes a bug in the ROM that computes OTP measurements to populate
the keymgr attestation binding CSRs, if the OTP fuse
(`OWNER_SW_CFG_ROM_KEYMGR_OTP_MEAS_EN`) to request this is set.
Additionally, this fixes the corresponding ROM E2E test as well, and
enhances its robustness.

Signed-off-by: Tim Trippel <ttrippel@google.com>