opentitanlib: accept ECDSA signatures as ASN.1 blobs
When processing ECDSA signatures consider the input file to be either
a raw signature, with 32 bytes R and S values in little endian format
concatenated into a single 64 bytes blob, or an ASN.1 encoded ECDSA
signature, which is a sequence of two big endian big number values,
which could be anywhere from one to 33 bytes long.

In case input is in ASN .1 format Convert the values into little
endian and truncate/expand as necessary to generated the internal
representation of R and S of exactly 32 bytes each.

Tested by attaching the Crypto4a HSM produced ECDSA signature to an
owner binary image and verifying the signature, both using
opentitantool.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>