[rom] Update clkmgr jitter enable to match ROM specification.
The ROM e2e plan captures the expected behavior of the clock jitter
enable flow:

> `CREATOR_SW_CFG_JITTER_EN`: `kMultiBitBool4False` disables, all other values
>  enables. `CREATOR_SW_CFG_AST_INIT_EN` must also be `kMultiBitBool4True` to enable.

This change refactors the code in `rom_start.S` and `test_rom_start.S`
to explicitly reduce any value different than MubiFalse to MubiTrue.
The implementation relies on the register to be reset to MubiFalse by
default.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>