[prim,clkmgr,rtl] Simplify prim_clock_meas by removing RefCnt
This parameter is only ever set to 1: we don't do things like testing
a clock that should be 50% faster by setting Cnt=30, RefCnt=20.
Simplify the code slightly by getting rid of the support.

This commit also expands a comment explaining Cnt: my initial read
suggested that this was the maximum value of a counter, which gives
the wrong width if Cnt is a power of two.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>