[orchestrator] fix confirmation requirement bug
Confirmation requirement should only be required when the
`--non-interactive` mode CLI arg is NOT passed. This updates the script
behavior to disble the confirmation requirement when said flag is passed
and a host/device device ID mismatch occurs, which typically happens on
parts that have had CP run, and later FT run, as opposed to running both
stages back to back.

Signed-off-by: Tim Trippel <ttrippel@google.com>