[prim_xilinx,englishbreakfast] Remove ROM dependency
Create prim_xilinx_rom that instantiates an xpm_memory_sprom when there
is no memory initialization file provided, and ensure the macro is
configured to remain after optimizations, so splicing is available.

Change the englishbreakfast build to start with a ROM that is all zeroes
and requires splicing to get a complete image.

Signed-off-by: Alexander Williams <awill@opentitan.org>