[rules] Rename and change semantiv of `extra_bazel_features`
The `extra_bazel_features` attribute was added #25840 as an attribute
to `rv_rule` to transitively add features to the rule. However this is
too general because when applied to opentitan_binary, it will change
the features not only for the rule, but all its dependencies such as
exec_env, tools,  etc which may have unintended effects.

This commit keeps essentially the same logic but instead of applying
it to `rv_rule`, it is now only an attribute of `opentitan_binary`
and `opentitan_test` and it transitions only features for the `deps`.
This attribute is also renamed to `transitive_features`. The rationale
for the rename is that every bazel rule has an implicit `features`
argument but it only applies to the target, not its dependencies. This
attribute does the same but also applies transitively.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>