[pinmux] Correct computation of WKUP_DETECTOR_PADSEL width
The `dio_in_i` signal going into the wakeup detector multiplexer is of
width `NDioPads`, i.e., it contains both inputs and outputs. Previously,
ipgen was using the number of DIO input pads to compute the width of
the padsel register fields. As a result, only a subset of the DIO pads
could be selected in Darjeeling.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>