[prim,rtl] Rewrite a case statement to avoid unreachable case
Instead of the default case being unreachable (unless state_q is X),
rewrite things so that the last item is caught by the default case.

This is clearly safe at the moment (there are just two possible
states!) but this adds the GoodState_A assertion to make sure no
future RTL change can make this unsafe.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>