[bazel,python] vendor in requirements.bzl
Our current version of rules_python autogenerates the `requirements.bzl`
file from a `requirements.txt` lock file. However, this is awkward if
this project is ever a dependency of another downstream bazel project,
as shown here: https://github.com/bazelbuild/rules_python/issues/608

To circumvent this issue, rules_python provides a mechanism to vendor in
the requirements.bzl file into the repo, which this commit does.

Signed-off-by: Tim Trippel <ttrippel@google.com>