[dv,darjeeling] Temporarily use Earl Grey ROM scrambling
The build flow presently creates ROM images only using Earl Grey
scrambling keys (read from top_earlgrey.gen.hjson), so in order
to unscramble the ROM images successfully and execute software,
temporarily modify the templated top_level for Darjeeling only
to use fixed keys.

Note that this only works for the standard build seed that we
employ, and that this is a temporary measure to avoid conflicting
with/replicating the work on multitop support in sw/building.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>