[opentitantool] Remove pub_key_id comment
Emitting the public key id comment is causing opentitantool to crash
during `image manifest show`.  This is related to eliminating the
dependency on `min_specialization` in serde_annotate, but I need to
re-examine the approach in that library.  To avoid min_specialization,
I took the approach of managing a raw pointer, but serde occasionally
will hand you a &&T when you were expecting &T (this isn't a problem in
serde because it implements a forwarding implementation of `Serialize
for &T where T: Serialize`).  When trying to manage a raw pointer, we
don't know whether we have `&T` or `&&T`; when we have the later,
attempting to access the data causes a crash.

Signed-off-by: Chris Frantz <cfrantz@google.com>