[python] switch from pip-tools to uv
This is a manual cherry-pick of 56766397c71221be93cf47524d031f8f27884737
to the `earlgrey_1.0.0` branch. This was originally part of #24521.

`uv pip compile` is compatible with `pip-tools`, but much faster.
Switch to uv also make us no longer have to depend on `pip`, which is
considered as an unsafe dependency by `pip-tools`.

Signed-off-by: Tim Trippel <ttrippel@google.com>