[rules] Add compatibility check about exec_env in binary deps
Currently opentitan_binary will happily consume a binary dependency
from an incompatible exec_env without realizing. This is because
it will only check for the provider of the expected exec_env and if
not present, it will default to picking the first available file.

This commit adds a new OpenTitanBinaryInfo provider to any target
built by opentitan_binary. This means that one can always detect
if a target was built by opentitan_binary and detect if no
compatible exec_env were found.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>