[sw,cert] Change the semantic of Integer and ByteArray types.
After this commit, the "Integer" is for variables encoded to DER integers with
different lengths depending on its content.

OTOH, the "ByteArray" is for variables that are either treated as bytes,
or will be encoded as an DER integer with its MSb set.

The only exception is the public keys, which are integer variables, but encoded
as padded integers. Changing them to byte-array requires non-trival refactoring
of the testing code which assumes the public key typed as integer.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>
Change-Id: I91a932d7a27dd16579a04136c412fd3261f34a5e