[tl,dv] Remove `chk_prot_addr_mask_align
This check wasn't actually doing anything. The code actually checks
that the inactive byte lanes for the operation (based on a_addr and
a_size) are all zero.

But that test is also in chk_prot_mask_in_enabled_lanes: no need to
double up!

The TileLink spec does actually have some extra requirements on
a_mask, which depend on the operation.

  - For PutFullData, we already check that the mask
    is full (using mask_is_full).

  - For PutPartialData, there aren't any extra requirements.

  - For Get, the mask is required to be "contiguous and naturally
    aligned". We actually check that for all opcodes (using the
    mask_contiguous_c constraint). Technically, this is stricter than
    required by PutPartialData. We don't do any natural alignment
    checks.

The questionable checking for Get operations is unchanged by this
commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>