[alert_handler, dv] Alert happened after ping
When ping happens we go to ping handshake phases and
meanwhile if alert happens we can't proceed with
alert handshake phases until ping finishes. But in
cip_base_scb, if a fatal alert happens then the task
check_alert_triggered() waits for 6 cycles to get the
alert in the middle of alert handshake.
This PR tweaks check_alert_triggered() to watch ping
handshakes in both of the phases and once the handshakes
finishes then wait for 6 cycles to get the alert in the
middle of alert handshake.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>