[rust] Remove `num_enum` dependency
The macros for `num_enum` try to find and parse the crate's Cargo.toml
which we don't have because we're using Bazel. This leads to a lot of
spam when compiling.

`strum` does most of what we need `num_enum` for anyway, so it has been
replaced and can now be removed.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>