[dtgen] Fix generator when a module is not accessible
Some modules may not be accessible the hart address space (e.g.
racl_ctrl on Darjeeling). In this case, the generator was not
creating an entry in the DT which is incorrect. With this commit,
an entry is always created with a purposefully wrong address
(0xffffffff). Though this is not ideal, it at least ensures that
the code compiles.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>