[aes,dv] Modify c_dpi_aes_crypt_message interface
This commit adds the following parameters to the
c_dpi_aes_crypt_message interface:
- Message length
- AAD Length
- Error result code of the underlying library

Previously, the first two parameters (length) were generate
by simply using the size() of the input vector. However,
as this input vector is always a multiple of 16 bytes,
sub-blocks smaller than 16 bytes are not supported.
As we need this for the AES-GCM mode, add a dedicated
length parameter to the interface.

Moreover, this commit also enables that the
c_dpi_aes_crypt_message function returns the error code
that is generated by the underyling crypto lib.

This is important as, when in GCM mode, a tag mismatch during
decryption is signaled using this error code.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>