[alert_handler, dv] Ping happened after alert
In alert_monitor we got ping_thread() and alert_thread()
running together in parallel. If alert happened and it
stays for multiple cycles even after completing itself
and meanwhile a ping happens and we are in the middle of
ping handshake then the chances are higher to fail checks
in cip_base_scoreboard. Because maybe ping can timeout
as it didn't receive ack or it is in the middle of its
handshake phases and scoreboard doesn't knows about that
(#26138).
For this particular PR, we wait in the ping_thread for
existing alert completion if it happens before ping. Due
to that won't get ourselves into the ping wait timeout
thread that waits to finish the ping handshake in 32
cycles.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>