[hw/top] Introduce Bzlmod extension to generate the top description
This extension essentially generates the same description as topgen
but it does not to check in any `defs.bzl` files in the repository.
On the technical side, the bzlmod extension create a repository
`tops_desc` which contains one description file per top + a `defs.bzl`
that pulls everything together. The files are generated by a python
script which is run by the repository rule.

There are a few tricky details in the python script. The first one
is that contrary to topgen, the script has to figure out from the
top Hjson where the IP's hjson files are. This is done by providing
the script a list of directories to search. By default, the repo rule
will include:
- <top dir>/ip
- <top dir>/ip_autogen
- <repo top>/hw/ip

In the future, extra directories could be added by adding attributes
to the extension to support out of tree IPs if necessary. The second
detail is that the script works with filenames but has to convert
those to bazel labels when generating the file, e.g.
  hw/ip/uart/data/uart.hjson -> //hw/ip/uart:data/uart.hsjon
This is done by passing the script a mapping from file system paths
to bazel repo names, e.g.
  <absolute path to repo> -> @//
This mechanism is quite flexible so that in the future, it's possible
to register out of tree tops, or tops in other repositories and it
will correct work. Obviously there is still some amout of guesswork
required, in particular if the hjson files are not visible/exported
by some BUILD files, the script will work but bazel will complain
afterwards about the top description.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>